restructured folders and updated imports
