+++ src/app/blog/[slug]/page.tsx
@@ -1,7 +1,7 @@
 import { notFound } from 'next/navigation'
-import { CustomMDX } from 'app/components/mdx'
-import { formatDate, getBlogPosts } from 'app/blog/utils'
-import { baseUrl } from 'app/sitemap'
+import { CustomMDX } from 'src/app/components/mdx'
+import { formatDate, getBlogPosts } from 'src/app/blog/utils'
+import { baseUrl } from 'src/app/lib/sitemap'
 
 export async function generateStaticParams() {
   let posts = getBlogPosts()
