+++ src/app/blog/posts/static-typing.mdx
